moved images and added color and image namespaces
