+++ src/PuzzleSolver.App/frmMain.cs
@@ -1,3 +1,4 @@
+using PuzzleSolver.Image;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
