+++ src/PuzzleSolver/Image/ImageStats.cs
@@ -1,10 +1,11 @@
-﻿using SixLabors.ImageSharp;
+﻿using PuzzleSolver.Color;
+using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace PuzzleSolver
+namespace PuzzleSolver.Image
 {
     public class ImageStats
     {
